feat(runtime): integrate resource lifecycle, render scalability and SDK updates
Checkpoint current project work including UI authoring, JSON/localization, U32 meshes, bounded frame-resource lookup and documentation. Runtime and render focused targets compile; broader validation follows in dedicated fixes.

+++ samples/3d_animation_graph/core/main.cpp
@@ -124,7 +124,7 @@ constexpr float QuadHalfHeightMeters = 0.25F;
     jointIndices.reserve(static_cast<usize>(VertexCount) * 4U);
     std::vector<u16> jointWeights;
     jointWeights.reserve(static_cast<usize>(VertexCount) * 4U);
-    std::vector<u16> indices;
+    std::vector<u32> indices;
     indices.reserve(IndexCount);
 
     for (u16 bone = 0; bone < BoneCount; ++bone)
@@ -151,10 +151,9 @@ constexpr float QuadHalfHeightMeters = 0.25F;
                                 {Tina::AssetFormat::SkinnedMeshWire::WeightScale, 0, 0, 0});
         }
 
-        const u16 base = static_cast<u16>(bone * VerticesPerBone);
-        indices.insert(indices.end(), {base, static_cast<u16>(base + 1U),
-                                       static_cast<u16>(base + 2U), base,
-                                       static_cast<u16>(base + 2U), static_cast<u16>(base + 3U)});
+        const u32 base = static_cast<u32>(bone) * VerticesPerBone;
+        indices.insert(indices.end(), {base, base + 1U, base + 2U,
+                                       base, base + 2U, base + 3U});
     }
 
     const std::array<Tina::AssetFormat::StaticMeshSubmeshDesc, 1> submeshes{
